+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.28.1-0 / Unreleased
 * [Fixed] #848 Drag End Error with dnd5 extension
+* [Fixed] #850 ext-childcounter doesn't work with custom icons
 
 # 2.28.1 / 2018-03-19
   * [Fixed] #844 Fix RTL for ext-table